move account name validation test
