+++ internal/services/batch/validate/account_name_test.go
@@ -1,1 +1,29 @@
 package validate
+
+import "testing"
+
+func TestValidateBatchAccountName(t *testing.T) {
+	testCases := []struct {
+		input       string
+		shouldError bool
+	}{
+		{"ab", true},
+		{"ABC", true},
+		{"abc", false},
+		{"123456789012345678901234", false},
+		{"1234567890123456789012345", true},
+		{"abc12345", false},
+	}
+
+	for _, test := range testCases {
+		_, es := AccountName(test.input, "name")
+
+		if test.shouldError && len(es) == 0 {
+			t.Fatalf("Expected validating name %q to fail", test.input)
+		}
+
+		if !test.shouldError && len(es) > 1 {
+			t.Fatalf("Expected validating name %q to fail", test.input)
+		}
+	}
+}
